requirements.txt
# requirements_fixed.txt
# Fixed requirements file for OpenAI + Pinecone Voice AI Agent

# Core dependencies
aiohttp>=3.8.0
asyncio  # Part of standard library
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
flask>=2.3.0
simple-websocket>=0.10.0

# Audio processing
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.0
# audioop is part of Python standard library

# OpenAI integration
openai>=1.3.0
tiktoken>=0.5.0

# Pinecone vector database
pinecone>=5.0.0

# Google Cloud services
google-cloud-speech>=2.27.0
google-cloud-texttospeech>=2.16.0 
google-auth>=2.23.0
google-api-core>=2.11.0
grpcio>=1.56.0
protobuf>=4.25.0,<5.0.0

# Twilio integration
twilio>=8.8.0

# Document processing
PyMuPDF>=1.23.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Async support
aiofiles>=23.0.0

# Performance monitoring
psutil>=5.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0